*: fix data race in the IgnoreTruncate (#42449)
close pingcap/tidb#42341
