+++ executor/partitiontest/BUILD.bazel
@@ -1,0 +1,17 @@
+load("@io_bazel_rules_go//go:def.bzl", "go_test")
+
+go_test(
+    name = "partitiontest_test",
+    timeout = "short",
+    srcs = [
+        "main_test.go",
+        "partition_test.go",
+    ],
+    flaky = True,
+    shard_count = 4,
+    deps = [
+        "//testkit",
+        "@com_github_pingcap_failpoint//:failpoint",
+        "@com_github_stretchr_testify//require",
+    ],
+)
