+++ expression/builtin_op_test.go
@@ -70,11 +70,11 @@ func TestUnary(t *testing.T) {
 func TestLogicAnd(t *testing.T) {
 	ctx := createContext(t)
 	sc := ctx.GetSessionVars().StmtCtx
-	origin := sc.IgnoreTruncate
+	origin := sc.IgnoreTruncate.Load()
 	defer func() {
-		sc.IgnoreTruncate = origin
+		sc.IgnoreTruncate.Store(origin)
 	}()
-	sc.IgnoreTruncate = true
+	sc.IgnoreTruncate.Store(true)
 
 	cases := []struct {
 		args     []interface{}
@@ -243,11 +243,11 @@ func TestBitXor(t *testing.T) {
 func TestBitOr(t *testing.T) {
 	ctx := createContext(t)
 	sc := ctx.GetSessionVars().StmtCtx
-	origin := sc.IgnoreTruncate
+	origin := sc.IgnoreTruncate.Load()
 	defer func() {
-		sc.IgnoreTruncate = origin
+		sc.IgnoreTruncate.Store(origin)
 	}()
-	sc.IgnoreTruncate = true
+	sc.IgnoreTruncate.Store(true)
 
 	cases := []struct {
 		args     []interface{}
@@ -289,11 +289,11 @@ func TestBitOr(t *testing.T) {
 func TestLogicOr(t *testing.T) {
 	ctx := createContext(t)
 	sc := ctx.GetSessionVars().StmtCtx
-	origin := sc.IgnoreTruncate
+	origin := sc.IgnoreTruncate.Load()
 	defer func() {
-		sc.IgnoreTruncate = origin
+		sc.IgnoreTruncate.Store(origin)
 	}()
-	sc.IgnoreTruncate = true
+	sc.IgnoreTruncate.Store(true)
 
 	cases := []struct {
 		args     []interface{}
@@ -395,11 +395,11 @@ func TestBitAnd(t *testing.T) {
 func TestBitNeg(t *testing.T) {
 	ctx := createContext(t)
 	sc := ctx.GetSessionVars().StmtCtx
-	origin := sc.IgnoreTruncate
+	origin := sc.IgnoreTruncate.Load()
 	defer func() {
-		sc.IgnoreTruncate = origin
+		sc.IgnoreTruncate.Store(origin)
 	}()
-	sc.IgnoreTruncate = true
+	sc.IgnoreTruncate.Store(true)
 
 	cases := []struct {
 		args     []interface{}
@@ -441,11 +441,11 @@ func TestBitNeg(t *testing.T) {
 func TestUnaryNot(t *testing.T) {
 	ctx := createContext(t)
 	sc := ctx.GetSessionVars().StmtCtx
-	origin := sc.IgnoreTruncate
+	origin := sc.IgnoreTruncate.Load()
 	defer func() {
-		sc.IgnoreTruncate = origin
+		sc.IgnoreTruncate.Store(origin)
 	}()
-	sc.IgnoreTruncate = true
+	sc.IgnoreTruncate.Store(true)
 
 	cases := []struct {
 		args     []interface{}
@@ -495,11 +495,11 @@ func TestUnaryNot(t *testing.T) {
 func TestIsTrueOrFalse(t *testing.T) {
 	ctx := createContext(t)
 	sc := ctx.GetSessionVars().StmtCtx
-	origin := sc.IgnoreTruncate
+	origin := sc.IgnoreTruncate.Load()
 	defer func() {
-		sc.IgnoreTruncate = origin
+		sc.IgnoreTruncate.Store(origin)
 	}()
-	sc.IgnoreTruncate = true
+	sc.IgnoreTruncate.Store(true)
 
 	testCases := []struct {
 		args    []interface{}
@@ -602,11 +602,11 @@ func TestIsTrueOrFalse(t *testing.T) {
 func TestLogicXor(t *testing.T) {
 	ctx := createContext(t)
 	sc := ctx.GetSessionVars().StmtCtx
-	origin := sc.IgnoreTruncate
+	origin := sc.IgnoreTruncate.Load()
 	defer func() {
-		sc.IgnoreTruncate = origin
+		sc.IgnoreTruncate.Store(origin)
 	}()
-	sc.IgnoreTruncate = true
+	sc.IgnoreTruncate.Store(true)
 
 	cases := []struct {
 		args     []interface{}
