+++ expression/builtin_other_test.go
@@ -32,10 +32,10 @@ import (
 func TestBitCount(t *testing.T) {
 	ctx := createContext(t)
 	stmtCtx := ctx.GetSessionVars().StmtCtx
-	origin := stmtCtx.IgnoreTruncate
-	stmtCtx.IgnoreTruncate = true
+	origin := stmtCtx.IgnoreTruncate.Load()
+	stmtCtx.IgnoreTruncate.Store(true)
 	defer func() {
-		stmtCtx.IgnoreTruncate = origin
+		stmtCtx.IgnoreTruncate.Store(origin)
 	}()
 	fc := funcs[ast.BitCount]
 	var bitCountCases = []struct {
@@ -68,7 +68,7 @@ func TestBitCount(t *testing.T) {
 			continue
 		}
 		sc := new(stmtctx.StatementContext)
-		sc.IgnoreTruncate = true
+		sc.IgnoreTruncate.Store(true)
 		res, err := count.ToInt64(sc)
 		require.NoError(t, err)
 		require.Equal(t, test.count, res)
