+++ types/convert_test.go
@@ -252,7 +252,7 @@ func TestConvertType(t *testing.T) {
 	_, err = d.ToDecimal(sc)
 	require.Truef(t, terror.ErrorEqual(err, ErrTruncatedWrongVal), "err %v", err)
 
-	sc.IgnoreTruncate = true
+	sc.IgnoreTruncate.Store(true)
 	v, err = d.ToDecimal(sc)
 	require.NoError(t, err)
 	require.Equal(t, "0", v.(*MyDecimal).String())
@@ -474,7 +474,7 @@ func TestConvertToBinaryString(t *testing.T) {
 
 func testStrToInt(t *testing.T, str string, expect int64, truncateAsErr bool, expectErr error) {
 	sc := new(stmtctx.StatementContext)
-	sc.IgnoreTruncate = !truncateAsErr
+	sc.IgnoreTruncate.Store(!truncateAsErr)
 	val, err := StrToInt(sc, str, false)
 	if expectErr != nil {
 		require.Truef(t, terror.ErrorEqual(err, expectErr), "err %v", err)
@@ -486,7 +486,7 @@ func testStrToInt(t *testing.T, str string, expect int64, truncateAsErr bool, ex
 
 func testStrToUint(t *testing.T, str string, expect uint64, truncateAsErr bool, expectErr error) {
 	sc := new(stmtctx.StatementContext)
-	sc.IgnoreTruncate = !truncateAsErr
+	sc.IgnoreTruncate.Store(!truncateAsErr)
 	val, err := StrToUint(sc, str, false)
 	if expectErr != nil {
 		require.Truef(t, terror.ErrorEqual(err, expectErr), "err %v", err)
@@ -498,7 +498,7 @@ func testStrToUint(t *testing.T, str string, expect uint64, truncateAsErr bool, 
 
 func testStrToFloat(t *testing.T, str string, expect float64, truncateAsErr bool, expectErr error) {
 	sc := new(stmtctx.StatementContext)
-	sc.IgnoreTruncate = !truncateAsErr
+	sc.IgnoreTruncate.Store(!truncateAsErr)
 	val, err := StrToFloat(sc, str, false)
 	if expectErr != nil {
 		require.Truef(t, terror.ErrorEqual(err, expectErr), "err %v", err)
@@ -603,7 +603,7 @@ func accept(t *testing.T, tp byte, value interface{}, unsigned bool, expected st
 	d := NewDatum(value)
 	sc := new(stmtctx.StatementContext)
 	sc.TimeZone = time.UTC
-	sc.IgnoreTruncate = true
+	sc.IgnoreTruncate.Store(true)
 	casted, err := d.ConvertTo(sc, ft)
 	require.NoErrorf(t, err, "%v", ft)
 	if casted.IsNull() {
